backend/requirements.txt
Flask
Flask-CORS
python-dotenv
bcrypt
werkzeug
gunicorn
firebase-admin
PyJWT
google-cloud-firestore